SCUMM: Split file.cpp to avoid referencing ScummEngine in detection code
This creates a specific translation unit for the code making use of
ScummEngine while keeping file.cpp for the classes needed for the
detection code.

This allows to build with UBSan while the Scumm engine is disabled.
